refactor(u01): extract shared sidecar connection core (PR2 — finale)
Completes U-01: the duplicated/divergent sidecar connection subsystem is
extracted into a single neutral module, shared/sidecarCore.ts, imported by
both tauri-app (Explorer) and tauri-prep (Constituer). Each client keeps only
its app-specific HTTP *endpoints* and re-exports the shared connection API so
existing `from "./sidecarClient"` imports keep working.

What moved to shared/sidecarCore.ts (canonical = prep, the version pinned by
T-05): module connection state (_conn/_connDbPath/_spawnedChild), Conn,
SidecarError, the SIDECAR_* constants, transport (sidecarFetch, makeConn,
_rawGet/_rawPost/_rawPut), lifecycle (ensureRunning, _spawnSidecar,
_readFirstJsonFromCommand, _notifyRustRegistry, _persistSidecarPort, pollHealth,
shutdownSidecar, getActiveConn, resetConnection), portfile + URL helpers, and
the shared utils (sidecarLog [neutral prefix now], errToString, sleep, …).

Result:
  - app's connection code DELETED (was its divergent fork) — it now runs prep's
    canonical, T-05-pinned connection logic. tauri-app/sidecarClient 1430→~570 l.
  - prep's connection code DELETED too — single source in shared. 2984→~2080 l.
  - In the shell, both modules import the SAME shared/sidecarCore → ONE
    sidecarCore chunk + ONE shared _conn (was two divergent caches). The two
    remaining sidecarClient chunks are the genuinely-different endpoint sets.
  - Explorer-alone stays possible: shared/ is self-contained (no @tauri-apps
    endpoint types, no prep dependency).

Wiring: each app's tsconfig gains `../shared` in include + a baseUrl/paths
mapping for @tauri-apps/* (tsc), and a vite resolve.alias for @tauri-apps/*
(rolldown can't resolve them from shared/ on its own). Dead @tauri-apps imports
removed from both clients.

Verified: tauri-prep/app/shell all tsc + build + vitest green (419/14/28); the
T-05 connection suite (25 tests) exercises the shared core via prep's
re-export; connection fns now exist ONLY in shared (app=prep=0).
