+++ tauri-app/vite.config.ts
@@ -1,3 +1,4 @@
+import { fileURLToPath, URL } from "node:url";
 import { defineConfig } from "vite";
 
 // https://vitejs.dev/config/
@@ -6,6 +7,16 @@ export default defineConfig(async () => ({
   //
   // 1. prevent vite from obscuring rust errors
   clearScreen: false,
+  resolve: {
+    // ../shared/sidecarCore lives outside this app's tree and can't resolve
+    // @tauri-apps/* on its own — point them at this app's deps.
+    alias: [
+      {
+        find: /^@tauri-apps\//,
+        replacement: fileURLToPath(new URL("./node_modules/@tauri-apps/", import.meta.url)),
+      },
+    ],
+  },
   // 2. tauri expects a fixed port, fail if that port is not available
   server: {
     port: 1420,
