+++ tauri-prep/tsconfig.json
@@ -3,7 +3,11 @@
     "target": "ES2022",
     "useDefineForClassFields": true,
     "module": "ESNext",
-    "lib": ["ES2022", "DOM", "DOM.Iterable"],
+    "lib": [
+      "ES2022",
+      "DOM",
+      "DOM.Iterable"
+    ],
     "skipLibCheck": true,
     "moduleResolution": "bundler",
     "allowImportingTsExtensions": true,
@@ -13,8 +17,19 @@
     "strict": true,
     "noUnusedLocals": false,
     "noUnusedParameters": false,
-    "noFallthroughCasesInSwitch": true
+    "noFallthroughCasesInSwitch": true,
+    "baseUrl": ".",
+    "paths": {
+      "@tauri-apps/*": [
+        "node_modules/@tauri-apps/*"
+      ]
+    }
   },
-  "include": ["src"],
-  "exclude": ["src/**/__tests__/**"]
+  "include": [
+    "src",
+    "../shared"
+  ],
+  "exclude": [
+    "src/**/__tests__/**"
+  ]
 }
