+++ tauri-prep/vite.config.ts
@@ -1,7 +1,18 @@
+import { fileURLToPath, URL } from "node:url";
 import { defineConfig } from "vite";
 
 export default defineConfig(async () => ({
   clearScreen: false,
+  resolve: {
+    // The shared connection core (../shared) lives outside this app's tree and
+    // can't resolve @tauri-apps/* on its own — point them at this app's deps.
+    alias: [
+      {
+        find: /^@tauri-apps\//,
+        replacement: fileURLToPath(new URL("./node_modules/@tauri-apps/", import.meta.url)),
+      },
+    ],
+  },
   server: {
     port: 1421,
     strictPort: true,
