+++ tauri-shell/tsconfig.json
@@ -3,7 +3,11 @@
     "target": "ES2020",
     "useDefineForClassFields": true,
     "module": "ESNext",
-    "lib": ["ES2020", "DOM", "DOM.Iterable"],
+    "lib": [
+      "ES2020",
+      "DOM",
+      "DOM.Iterable"
+    ],
     "skipLibCheck": true,
     "moduleResolution": "bundler",
     "allowImportingTsExtensions": true,
@@ -13,7 +17,18 @@
     "strict": true,
     "noUnusedLocals": false,
     "noUnusedParameters": false,
-    "noFallthroughCasesInSwitch": true
+    "noFallthroughCasesInSwitch": true,
+    "baseUrl": ".",
+    "paths": {
+      "@tauri-apps/*": [
+        "node_modules/@tauri-apps/*"
+      ]
+    }
   },
-  "include": ["src", "../tauri-prep/src", "../tauri-app/src"]
+  "include": [
+    "src",
+    "../tauri-prep/src",
+    "../tauri-app/src",
+    "../shared"
+  ]
 }
