+++ tauri-shell/vite.config.ts
@@ -1,7 +1,18 @@
+import { fileURLToPath, URL } from "node:url";
 import { defineConfig } from "vite";
 
 export default defineConfig(async () => ({
   clearScreen: false,
+  resolve: {
+    // ../shared/sidecarCore (imported via the app/prep modules) can't resolve
+    // @tauri-apps/* on its own — point them at the shell's deps.
+    alias: [
+      {
+        find: /^@tauri-apps\//,
+        replacement: fileURLToPath(new URL("./node_modules/@tauri-apps/", import.meta.url)),
+      },
+    ],
+  },
   server: {
     port: 1422,
     strictPort: true,
